formulario_culsen/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdfplumber==0.10.3
PyPDF2==3.0.1
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-api-python-client==2.107.0
gspread==5.12.0
yagmail==0.15.293
keyring==24.2.0
anthropic==0.5.0
openai==1.3.5
google-generativeai==0.3.1
python-dateutil==2.8.2
requests==2.31.0
python-dotenv==1.0.0 